# src/main/java/org/coursekata/controller/AuthController.java
package org.coursekata.controller;

import lombok.extern.log4j.Log4j2;
import org.apache.logging.log4j.message.StringMapMessage;
import org.coursekata.exception.TokenRefreshException;
import org.coursekata.model.auth.AuthenticationResponse;
import org.coursekata.model.auth.TokenRefreshRequest;
import org.coursekata.service.AuthService;
import org.coursekata.utils.HttpHeaderUtils;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpHeaders;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@Log4j2
@RestController
@RequestMapping("/auth")
public class AuthController {

    private static final String MESSAGE_KEY = "Message";
    private static final String TOKEN_KEY = "Token";

    private final AuthService authService;

    @Autowired
    public AuthController(AuthService authService) {
        this.authService = authService;
    }

    @GetMapping("/issue")
    public ResponseEntity<AuthenticationResponse> issueToken() {
        logInfo();

        AuthenticationResponse authenticationResponse = authService.generateInitialTokenResponse();
        HttpHeaders headers = HttpHeaderUtils.createAuthorizationHeader(authenticationResponse.getToken());

        logInfo("Initial token issued successfully", TOKEN_KEY, authenticationResponse.getToken());
        return ResponseEntity.ok().headers(headers).body(authenticationResponse);
    }

    @PostMapping("/refresh")
    public ResponseEntity<AuthenticationResponse> refreshToken(@RequestBody TokenRefreshRequest refreshRequest) {
        logInfo("Received request to refresh JWT token", TOKEN_KEY, refreshRequest.getToken());

        try {
            AuthenticationResponse authenticationResponse = authService.refreshTokenResponse(refreshRequest.getToken());
            logInfo("Token refreshed successfully", "NewToken", authenticationResponse.getToken());
            return ResponseEntity.ok(authenticationResponse);
        } catch (TokenRefreshException e) {
            logError(refreshRequest.getToken(), e);
            throw e;
        }
    }

    private void logInfo() {
        log.info(new StringMapMessage().with(MESSAGE_KEY, "Issuing initial JWT token for session"));
    }

    private void logInfo(String message, String key, String value) {
        log.info(new StringMapMessage().with(MESSAGE_KEY, message).with(key, value));
    }

    private void logError(String value, Exception e) {
        log.error(new StringMapMessage().with(MESSAGE_KEY, "Error refreshing token").with(
            AuthController.TOKEN_KEY, value), e);
    }
}